requirements.txt

requests
sqlite3
requests
python-dotenv
cryptography